// frontend/src/app/api/model/userRegistrationDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserRegistrationDto { 
    fname?: string;
    lname?: string;
    email?: string;
    password?: string;
    bdate?: string;
    lvl?: string;
    siteId?: number;
    roleId?: number;
}